+++ cpp/cluster/ClusterProperties.cc
@@ -18,10 +18,11 @@ namespace freud { namespace cluster {
     compute loops over all points in the given array and determines the center
     of mass of the cluster as well as the gyration tensor. These can be
     accessed after the call to compute with getClusterCenters() and
-    getClusterGyrations().
+    getClusterInertiaMoments().
 */
 
-void ClusterProperties::compute(const freud::locality::NeighborQuery* nq, const unsigned int* cluster_idx)
+void ClusterProperties::compute(const freud::locality::NeighborQuery* nq, const unsigned int* cluster_idx,
+                                const float* masses)
 {
     // determine the number of clusters
     const unsigned int* max_cluster_id = std::max_element(cluster_idx, cluster_idx + nq->getNPoints());
@@ -30,8 +31,9 @@ void ClusterProperties::compute(const freud::locality::NeighborQuery* nq, const 
     // allocate memory for the cluster properties and temporary arrays
     // initialize arrays to 0
     m_cluster_centers.prepare(num_clusters);
-    m_cluster_gyrations.prepare({num_clusters, 3, 3});
+    m_cluster_inertia_moments.prepare({num_clusters, 3, 3});
     m_cluster_sizes.prepare(num_clusters);
+    m_cluster_masses.prepare(num_clusters);
 
     // Create a vector to store cluster points, used to compute center of mass
     std::vector<std::vector<vec3<float>>> cluster_points(num_clusters, std::vector<vec3<float>>());
@@ -45,16 +47,19 @@ void ClusterProperties::compute(const freud::locality::NeighborQuery* nq, const 
         const unsigned int c = cluster_idx[i];
         cluster_points[c].push_back((*nq)[i]);
         m_cluster_sizes[c]++;
+        m_cluster_masses[c] += masses[i];
     }
 
     // Now that we have located all of the cluster vectors, compute the centers
+
     for (unsigned int c = 0; c < num_clusters; c++)
     {
-        m_cluster_centers[c] = nq->getBox().centerOfMass(cluster_points[c].data(), m_cluster_sizes[c]);
+        m_cluster_centers[c]
+            = nq->getBox().centerOfMass(cluster_points[c].data(), m_cluster_sizes[c], masses);
     }
 
     // Now that we have determined the centers of mass for each cluster, tally
-    // up the gyration tensor. This has to be done in a loop over the points.
+    // up the moment of inertia tensor. This has to be done in a loop over the points.
     for (unsigned int i = 0; i < nq->getNPoints(); i++)
     {
         unsigned int c = cluster_idx[i];
@@ -62,30 +67,30 @@ void ClusterProperties::compute(const freud::locality::NeighborQuery* nq, const 
         vec3<float> delta = nq->getBox().wrap(pos - m_cluster_centers[c]);
 
         // get the start pointer for our 3x3 matrix
-        m_cluster_gyrations(c, 0, 0) += delta.x * delta.x;
-        m_cluster_gyrations(c, 0, 1) += delta.x * delta.y;
-        m_cluster_gyrations(c, 0, 2) += delta.x * delta.z;
-        m_cluster_gyrations(c, 1, 0) += delta.y * delta.x;
-        m_cluster_gyrations(c, 1, 1) += delta.y * delta.y;
-        m_cluster_gyrations(c, 1, 2) += delta.y * delta.z;
-        m_cluster_gyrations(c, 2, 0) += delta.z * delta.x;
-        m_cluster_gyrations(c, 2, 1) += delta.z * delta.y;
-        m_cluster_gyrations(c, 2, 2) += delta.z * delta.z;
+        m_cluster_inertia_moments(c, 0, 0) += delta.x * delta.x * masses[i];
+        m_cluster_inertia_moments(c, 0, 1) += delta.x * delta.y * masses[i];
+        m_cluster_inertia_moments(c, 0, 2) += delta.x * delta.z * masses[i];
+        m_cluster_inertia_moments(c, 1, 0) += delta.y * delta.x * masses[i];
+        m_cluster_inertia_moments(c, 1, 1) += delta.y * delta.y * masses[i];
+        m_cluster_inertia_moments(c, 1, 2) += delta.y * delta.z * masses[i];
+        m_cluster_inertia_moments(c, 2, 0) += delta.z * delta.x * masses[i];
+        m_cluster_inertia_moments(c, 2, 1) += delta.z * delta.y * masses[i];
+        m_cluster_inertia_moments(c, 2, 2) += delta.z * delta.z * masses[i];
     }
 
-    // Normalize by the cluster size.
+    // Normalize by the cluster masses.
     for (unsigned int c = 0; c < num_clusters; c++)
     {
-        auto s = static_cast<float>(m_cluster_sizes[c]);
-        m_cluster_gyrations(c, 0, 0) /= s;
-        m_cluster_gyrations(c, 0, 1) /= s;
-        m_cluster_gyrations(c, 0, 2) /= s;
-        m_cluster_gyrations(c, 1, 0) /= s;
-        m_cluster_gyrations(c, 1, 1) /= s;
-        m_cluster_gyrations(c, 1, 2) /= s;
-        m_cluster_gyrations(c, 2, 0) /= s;
-        m_cluster_gyrations(c, 2, 1) /= s;
-        m_cluster_gyrations(c, 2, 2) /= s;
+        auto s = static_cast<float>(m_cluster_masses[c]);
+        m_cluster_inertia_moments(c, 0, 0) /= s;
+        m_cluster_inertia_moments(c, 0, 1) /= s;
+        m_cluster_inertia_moments(c, 0, 2) /= s;
+        m_cluster_inertia_moments(c, 1, 0) /= s;
+        m_cluster_inertia_moments(c, 1, 1) /= s;
+        m_cluster_inertia_moments(c, 1, 2) /= s;
+        m_cluster_inertia_moments(c, 2, 0) /= s;
+        m_cluster_inertia_moments(c, 2, 1) /= s;
+        m_cluster_inertia_moments(c, 2, 2) /= s;
     }
 }